% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePercentNotNull.R
\name{calcPercentNotNull}
\alias{calcPercentNotNull}
\title{Calculate percent not null}
\usage{
calcPercentNotNull(x)
}
\arguments{
\item{x}{A vector, matrix, or data frame.}
}
\value{
A single numeric value: the percentage of non-\code{NA} values in \code{x}.
}
\description{
Calculates the percentage of values in \code{x} that are not \code{NA}
(i.e., that carry valid SNP data). Used to summarise the proportion of loci
with usable data for a colony or a pairwise comparison.
}
