% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkforAllowableData.R
\name{checkforAllowableData}
\alias{checkforAllowableData}
\title{Filter valid base pair combinations}
\usage{
checkforAllowableData(x)
}
\arguments{
\item{x}{the dataset to use}
}
\description{
This function is used to select only only columns that contain allowable characters (those listed in IUPAC)
by the convertBasePairstoAlleles function. DOESNT THIS JUST RETURN THE ALLELEPAIRS?
}
