% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralesce-package.R
\docType{package}
\name{Coralesce-package}
\alias{Coralesce}
\alias{Coralesce-package}
\title{Coralesce: Group Coral Colonies Based on Genetic Relatedness}
\description{
Assigns coral colonies to clonal groups (genets) from single nucleotide polymorphism (SNP) data by comparing alleles at each locus across all pairwise combinations of colonies, and computes individual- and population-level mean kinship and gene diversity (excluding invariant loci). Wrapper functions read genotype CSV files from a 'Data' folder and write results to a 'Results' folder.
}
\author{
\strong{Maintainer}: Colin Shea \email{Colin.Shea@MyFWC.com}

Authors:
\itemize{
  \item Lauren Gentry \email{Lauren.Gentry@MyFWC.com}
  \item Kristen Sauby \email{Kristen.Sauby@MyFWC.com}
}

}
\keyword{internal}
