% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOnlyDataAllowed.R
\name{filterOnlyDataAllowed}
\alias{filterOnlyDataAllowed}
\title{Filter valid base pair combinations}
\usage{
filterOnlyDataAllowed(x)
}
\arguments{
\item{x}{the dataset to use}
}
\description{
This function removes invalid base pair combinations and is used
by the convertBasePairstoAlleles function. DOESNT THIS JUST RETURN THE ALLELEPAIRS?
}
