% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecks.R
\name{find_dups}
\alias{find_dups}
\title{Notify user that at least one colony has multiple entries (i.e., rows) of SNP data, which is not permitted.}
\usage{
find_dups(df)
}
\arguments{
\item{dataset}{A data frame to be tested with a \code{Coral_ID} column and SNP data, along with other fields such as site name etc.}

\item{acceptableData}{A data frame with acceptable data that is allowable; here, these are \code{IUPAC} SNP data.}
}
\value{
This function returns a warning message notifying users of the presence and identity of columns in their data that do not adhere to formatting requirements. This is used in conjunction with \code{handleError_ProhibitedData()} in the \code{readGeneticData()}, which uses this function and also omits the offending columns.
}
\description{
This function notifies users that that at least one colony has multiple entries (i.e., rows) of SNP data, which is not permitted because downstream functions involve a pivot that requires unique row (colony) names. The process is stopped and the offending colonies are identified for the user.
}
