% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecks.R
\name{handleError_MatchMakerIndex}
\alias{handleError_MatchMakerIndex}
\title{Notify user that \code{MatchMaker_Index} field name had to be changed.}
\usage{
handleError_MatchMakerIndex(dataset)
}
\arguments{
\item{dataset}{A data frame with some sort of coral colony identifier (ideally named \code{MatchMaker_Index} and \code{Coral_ID}) and SNP genetic data (and perhaps other fields) in the remaining columns. The MatchMaker_Index is not used explicitly by MatchMakeR and is simply used for bookkeeping purposes in the database.}
}
\value{
This function returns a warning message notifying users that the MatchMakeR colony identifier field was changed to \code{MatchMaker_Index} prior to any genet assignment or kinship calculations. If the column is already called \code{MatchMaker_Index} then nothing happens and no warning message is issued.
}
\description{
This function notifies users that the \code{MatchMaker_Index} column name was changed.
}
