% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecks.R
\name{handleError_ProhibitedData}
\alias{handleError_ProhibitedData}
\title{Notify user that non-conforming fields were found and omitted.}
\usage{
handleError_ProhibitedData(dataset, acceptableData)
}
\description{
This function notifies users that non-conforming fields/columns such as site name or, more importantly, allele pairs that are not included in the IUPAC reference file, were omitted from the data set. This function omits the offending columns and reports a message that they were removed along with their names.
}
