% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IUPAC}
\alias{IUPAC}
\title{IUPAC data}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{Allelepairs}{Character with two letters separated by a colon; ? for no data}
\item{IUPACallele}{One-letter character; ? for no data}
}
}
\usage{
IUPAC
}
\description{
A lookup table for converting 2-character alleles to single-letter IUPAC codes
}
\keyword{datasets}
