% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCloneStatus.R
\name{returnGenetIdentity}
\alias{returnGenetIdentity}
\title{For each coral pair observation, return the genet identity}
\usage{
returnGenetIdentity(obs)
}
\arguments{
\item{obs}{a dataframe of three columns named coral1, coral2, and CoralPair. coral1 identifies the first coral in the pair, coral2 identifies the second coral in the pair, and CoralPair identifies the pair, taking the format of the first coral identity concatenated together with the second coral identity, separated by a period.}
}
\description{
For each coral pair, return the identity of the genet to which the pair belongs..
}
\examples{
example <- data.frame(
  coral1=c(1,2,3,3,4,5,6),
  coral2=c(1,2,2,3,4,5,6),
  CoralPair=c("1.1","2.2","3.2","3.3","4.4","5.5","6.6")
)
returnGenetIdentity(example)
}
