% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePercentNotNull.R
\name{calcPercentNotNull}
\alias{calcPercentNotNull}
\title{Calculate percent not null}
\usage{
calcPercentNotNull(x)
}
\description{
This function calculates the percentage of columns (loci) that are NOT NA values i.e., that have valid SNP data.
}
