% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkforAllowableData.R
\name{checkforAllowableData}
\alias{checkforAllowableData}
\title{Filter valid base pair combinations}
\usage{
checkforAllowableData(dataset)
}
\arguments{
\item{dataset}{A data frame, where the first column titled "Coral_ID" uniquely identifies each row, and the second through last column each contain allele data for a single locus.}
}
\description{
This function is used to select only only columns that contain allowable characters (those listed in IUPAC).
}
