% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkforAllowableData.R
\name{checkforAllowableData}
\alias{checkforAllowableData}
\title{Filter valid base pair combinations}
\usage{
checkforAllowableData(dataset)
}
\arguments{
\item{dataset}{A data frame, where the first column named \code{Coral_ID} uniquely identifies the colony in each row, and the second through last column each contain allele data for a single locus.}
}
\value{
description
}
\description{
This function is used to only select columns that contain allowable characters; specifically, the allele pairs listed in the \code{IUPAC} table.
}
