% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkforAllowableData.R
\name{checkforAllowableData}
\alias{checkforAllowableData}
\title{Flag values that are valid IUPAC allele pairs}
\usage{
checkforAllowableData(x)
}
\arguments{
\item{x}{A vector of values to test (typically one column of the input data).}
}
\value{
A logical vector the same length as \code{x}, \code{TRUE} where the value is a
valid allele pair in \code{IUPAC$Allelepairs}.
}
\description{
Tests, element by element, whether the supplied values are
allowable allele pairs as listed in the \code{IUPAC} reference table. Used to
identify which columns of an input data set contain valid SNP data.
}
