% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyAllelePairs.R
\name{classifyAllelePairs}
\alias{classifyAllelePairs}
\title{Classify allele pairs}
\usage{
classifyAllelePairs(dataset, locus)
}
\arguments{
\item{dataset}{A data frame, where the first column titled "Coral_ID" uniquely identifies each row, and the second through last column each contain allele data for a single locus.}

\item{locus}{The column number containing allele data for a single locus}
}
\description{
This function creates a data frame of all possible pairwise combinations of individuals and determines, for each locus, whether two individuals have the same allele or not. Comparisons are made for each colony with themselves AND with all other colonies. This function works on one locus (i.e., dataset column) at a time.
}
