% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyAllelePairs.R
\name{classifyAllelePairs}
\alias{classifyAllelePairs}
\title{Classify allele pairs at a single locus for all pairwise combinations of individuals}
\usage{
classifyAllelePairs(dataset, locus)
}
\arguments{
\item{dataset}{A dataframe, where the first column titled "Coral_ID" uniquely identifies each row, and the second through last column each contain allele data for a single locus.}

\item{locus}{the column number containing allele data for a single locus}
}
\description{
Create a data frame of all possible pairwise combinations of individuals and determine for a single locus whether each pair of individuals have the same allele or not.
}
