% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyAllelePairsOthers.R
\name{classifyAllelePairsOthers}
\alias{classifyAllelePairsOthers}
\title{Classify allele pairs}
\usage{
classifyAllelePairsOthers(dataset, locus)
}
\arguments{
\item{dataset}{A data frame, where the first column titled "Coral_ID" uniquely identifies each row, and the second through last column each contain allele data for a single locus.}

\item{locus}{The column number containing allele data for a single locus}
}
\description{
This function creates a data frame of all possible pairwise combinations of individuals and determines, for each  single locus, whether two individuals have the same allele or not. This function only returns comparisons for colonies with other colonies, and not comparisons of colonies with themselves (the classifyAllelePairs() function does both). This function works on one locus (i.e., dataset column) at a time.
}
