% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyAllelePairsOthers.R
\name{classifyAllelePairsOthers}
\alias{classifyAllelePairsOthers}
\title{Classify allele pairs}
\usage{
classifyAllelePairsOthers(dataset, locus)
}
\arguments{
\item{dataset}{A data frame, where the first column titled "Coral_ID" uniquely identifies each row, and the second through last column each contain allele data for a single locus.}

\item{locus}{The column number containing allele data for a single locus}
}
\description{
Create a data frame of all possible pairwise combinations of individuals and determine for a single locus whether each pair of individuals have the same allele or not. This function is only for comparisons with other colonies, not comparisons with others AND themelves (that's what classifyAllelePairs does).
}
