% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBasePairstoCodes.R
\name{convertBasePairstoCodes}
\alias{convertBasePairstoCodes}
\title{Convert base pairs to single character IUPAC codes}
\usage{
convertBasePairstoCodes(initdata)
}
\arguments{
\item{initdata}{A data frame containing a column "Coral_ID" with unique value per row, as well as at least one column containing base pair data with at least one value matching base pair data in IUPAC.}
}
\description{
This function replaces the base pairs with IUPAC codes and converts the data to wide format so that there is a column per locus. If Any invalid characters are present in a column, then the entire column is deleted, a warning is issued, and the column names are reported.
}
