% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBasePairstoCodes.R
\name{convertBasePairstoCodes}
\alias{convertBasePairstoCodes}
\title{Convert to long format and then convert base pairs to IUPAC codes}
\usage{
convertBasePairstoCodes(initdata)
}
\arguments{
\item{initdata}{A dataset containing a column "Coral_ID" with unique value per row, as well as at least one column containing base pair data with at least one value matching base pair data in IUPAC.}
}
\description{
This function replaces the base pairs with IUPAC codes and
converts the data to wide format so that there is a column per code.
}
\section{Make sure that the data we're using are ONLY either the}{
Coral_ID OR one of the character strings included in OnlyDataAllowed. If
there are no such columns then nothing happens. There's probably a more
elegant way to do this but it works fine.
}

