% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineAlleleMatches.R
\name{determineAllAlleleMatches}
\alias{determineAllAlleleMatches}
\title{Assess matching of alleles, comparing to selves and other colonies.}
\usage{
determineAllAlleleMatches(dataset)
}
\arguments{
\item{dataset}{A data frame, where the first column titled "Coral_ID" uniquely identifies each row, and the second through last column each contain allele data for a single locus.}
}
\description{
For each pairwise combination of individuals, this function classifies allele pairs at each locus as either matching or not. This function applies the classifyAllelePairs function, one column or locus at a time, by looping sequentially through all of N loci. This function includes allele matches at each locus for colonies compared to themselves and compared to other colonies.
}
