% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineAlleleMatchesOthers.R
\name{determineAllAlleleMatchesOthers}
\alias{determineAllAlleleMatchesOthers}
\title{Classify allele pairs at all loci for all pairwise combinations of individuals, but exclude comparisons with selves}
\usage{
determineAllAlleleMatchesOthers(dataset)
}
\arguments{
\item{dataset}{A dataframe, where the first column titled "Coral_ID" uniquely identifies each row, and the second through last column each contain allele data for a single locus.}
}
\description{
For each pairwise combination of individuals, classify allele pairs at each locus as either matching or not. This works one column (locus) at a time, by looping through all loci
}
