% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCloneStatus.R
\name{groupByGenets}
\alias{groupByGenets}
\title{Determine if the two individuals in the pairwise comparison come from the
same genet}
\usage{
groupByGenets(
  CoralAlleleData,
  AlleleMatchResults,
  PctMatchThreshold = NULL,
  PctNotNullThreshold = NULL
)
}
\arguments{
\item{PctMatchThreshold}{Defaults to 99.}

\item{finalResultsWide}{the results to return in "wide" format}
}
\description{
The function calculates the percentage match and percentage not
null and uses this information to determine if the pairwise comparison
indicates that the two individuals are from the same genet.
}
