% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupByGenets.R
\name{groupByGenets}
\alias{groupByGenets}
\title{Determine if individuals come from the same genet.}
\usage{
groupByGenets(
  CoralAlleleData,
  AlleleMatchResults,
  PctMatchThreshold = NULL,
  PctNotNullThreshold = NULL,
  getPairwiseAlleleMatches = FALSE
)
}
\arguments{
\item{CoralAlleleData}{Raw coral data (rows = coral colonies and columns = loci; with one-letter codes for allele combinations and \code{NA} for \verb{?} or blank values). This is used to append the colony-specific \code{pctNull} values to the genet assignment file created by this function.}

\item{AlleleMatchResults}{data set with results of pairwise matching of alleles at each locus for all individuals and loci}

\item{PctMatchThreshold}{Defaults to NULL but must be specified (can be specified in \code{runGenets} wrapper function)}

\item{PctNotNullThreshold}{Defaults to NULL but must be specified (can be specified in \code{runGenets} wrapper function)}
}
\description{
The function calculates the percentage match and percentage not null and uses this information to determine if the pairwise comparison indicates that the two individuals are from the same genet. Uses \code{returnGenetIdentity} for the genet classification. #' @importFrom dplyr arrange if_else n rename add_row distinct
}
