% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecks.R
\name{handleError_allZeros}
\alias{handleError_allZeros}
\title{Notify user that colonies with no valid SNP data were found.}
\usage{
handleError_allZeros(dataset)
}
\description{
This function finds colonies with all NA values, report with a message, and separates the \code{allNA} individuals from the rest of the data. These \verb{all NA}individuals are appended to the genet classification file and assigned genet = NA, pctNULL = 100, and adequateData = No.
}
