% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecks.R
\name{handleError_allZeros}
\alias{handleError_allZeros}
\title{Notify user that colonies with no valid SNP data were found.}
\usage{
handleError_allZeros(dataset)
}
\arguments{
\item{dataset}{Takes a file with SNP data with a Coral_ID column as the first column and SNP data in the remaining columns.}
}
\value{
Returns a data frame of colonies with all NA values. If this is NULL, then nothing happens, but if this is a data frame with at least one observation, then isolateAllNAColonies() will remove the offending individuals and append to the genet assigment table.
}
\description{
This function finds colonies with all NA values, report with a message, and separates the \code{allNA} individuals from the rest of the data. These \verb{all NA}individuals are appended to the genet classification file and assigned genet = NA, pctNULL = 100, and adequateData = No.
}
