% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecks.R
\name{handleError_CoralID}
\alias{handleError_CoralID}
\title{Notify user that Coral_ID field name had to be changed.}
\usage{
handleError_CoralID(dataset)
}
\arguments{
\item{dataset}{A data frame with some sort of coral colony identifier (ideally named Coral_ID)}
}
\description{
This function notifies users that the Coral_ID column name was changed.
}
