% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecks.R
\name{handleError_ProhibitedData}
\alias{handleError_ProhibitedData}
\title{Notify user that non-conforming fields were omitted from the data set}
\usage{
handleError_ProhibitedData(dataset, acceptableData)
}
\description{
Notify user that non-conforming fields were omitted from the data set
}
