% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolateAllNAColonies.R
\name{isolateAllNAColonies}
\alias{isolateAllNAColonies}
\title{Isolate all NA colonies}
\usage{
isolateAllNAColonies(dataset)
}
\arguments{
\item{dataset}{A dataset containing a column "Coral_ID" base pair data at each of N loci.}
}
\value{
A data frame called allNA and version of the input data frame that excludes the allNA colonies. If there are no allNA colonies then the original data frame is returned and the allNA data frame is set to NULL.
}
\description{
This function determines whether any colonies have NA values at all loci, in which case they are isolated in a new data frame called allNA. The remaining data frame contains colonies with at an allele at at least one locus.
}
