% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolateAllNAColonies.R
\name{isolateAllNAColonies}
\alias{isolateAllNAColonies}
\title{Isolate all NA colonies}
\usage{
isolateAllNAColonies(dataset)
}
\arguments{
\item{dataset}{A data frame containing a \code{Coral_ID} column for each of N colonies and SNP data at each of N loci.}
}
\value{
A data frame called \code{allNA} and version of the input data frame that excludes the \code{allNA} colonies. If there are no \code{allNA} colonies then the original data frame is returned and the \code{allNA} data frame is set to \code{NULL}.
}
\description{
This function determines whether any colonies have NA values at all loci, in which case they are isolated in a new data frame called allNA. The remaining data frame contains colonies with SNP data at, at the very least, one locus.
}
