% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolateAllNAColonies.R
\name{isolateAllNAColonies}
\alias{isolateAllNAColonies}
\title{Isolate all NA colonies}
\usage{
isolateAllNAColonies(dataset)
}
\arguments{
\item{dataset}{A data frame containing a \code{Coral_ID} column for each of N colonies and valid \code{IUPAC-allele} codes (i.e., G, not G:G) for SNP data at each of N loci. This function is used in conjunction with \code{convertBasePairstoCodes()} during initial processing.}
}
\value{
This function returns two objects: (1) if such individuals are present, a data frame called \code{allNA} is returned with the identify of \code{allNA} colonies, and (2) a version of the input data frame that excludes any \code{allNA} colonies. If there are no \code{allNA} colonies, then the original data frame is returned and the \code{allNA} data frame is a \code{NULL} object. Additionally, if \code{allNA} individuals are present, their identifies will be listed in a warning message issues by the \code{handleError_allZeros()} function.
}
\description{
This function determines whether any colonies have NA values at all loci, in which case they are isolated in a new data frame called allNA. The remaining data frame contains colonies with SNP data at, at the very least, one locus.
}
