% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinshipCalcs.R
\name{kinshipCalcs}
\alias{kinshipCalcs}
\title{Calculate pairwise kinship for all possible pairwise comparisons.}
\usage{
kinshipCalcs(dataset)
}
\arguments{
\item{dataset}{A data frame with Coral_ID column as unique identifier (rows) and loci as columns; this is a data frame resulting from the \code{convertBasePairstoCodes()} function.}
}
\value{
This function returns data frame with the calculations is returned. This function calculates pairwise kinship for all possible comparisons of alleles among individuals. Note that these are simply calculation and this function does not exclude invariant loci (another function used by called \code{omitInvariantLoci()} does that).
}
\description{
This function calculates pairwise kinship based on all alleles present at all possible combinations of colonies and loci.
}
