% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinshipCalcs.R
\name{kinshipCalcs}
\alias{kinshipCalcs}
\title{Calculate pairwise kinship for all pairwise comparisons across all individuals (selves and others) and loci (variant and invariant)}
\usage{
kinshipCalcs(dataset)
}
\arguments{
\item{dataset}{data resulting from determineAllAlleleMatches, which itself uses classifyAllelePairs}
}
\description{
Calculates pairwise kinship across all individuals and loci and does not exclude invariant loci
}
