% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinshipCalcs.R
\name{kinshipCalcs}
\alias{kinshipCalcs}
\title{Calculate pairwise kinship for all possible pairwise comparisons.}
\usage{
kinshipCalcs(dataset)
}
\arguments{
\item{dataset}{A data frame containing...}
}
\description{
Calculates pairwise kinship based on all alleles present at all possible combinations of colonies and loci.
}
