% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinshipCalcsNoInvar.R
\name{kinshipCalcsNoInvar}
\alias{kinshipCalcsNoInvar}
\title{Calculate average individual-level and population-level kinship and gene diversity}
\usage{
kinshipCalcsNoInvar(dataset, targetN = NULL, subset = FALSE)
}
\arguments{
\item{dataset}{A data frame with Coral_ID column as unique identifier (rows) and loci as columns; this is a data frame resulting from the  \code{convertBasePairstoCodes} function.}

\item{targetN}{The desired number of individuals over which population-average kinship and gene diversity are calculated. This is ignored if left as NULL or if its value is greater than or equal to nrow(dataset) i.e., the number of individuals in a data set, nothing happens. If this value is smaller than nrow(), then kinship is recalculated repeatedly, removing the individual with the highest average kinship, one-by-one, until targetN individuals with the lowest averge kinship remain.}
}
\value{
This function returns three objects, PopAvgMKGD, MK_init, and MK_final. PopAvgMKGD contains the population-level average kinship and gene diversity; MK_init contains average kinship for each individual in dataset (i.e., now(dataset) individuals), and MK_final contains average kinship for the targetN individuals with the lowest average kinship.
}
\description{
Calculates average individual-level and population-level kinship and gene diversity (1 - kinship) across all individuals and loci, excluding invariant loci, based on all possible among-locus pairwise comparisons at each of N loci. Within-individual comparisons and invariant loci are excluded from these calculations.
}
