% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omitInvariantLoci.R
\name{omitInvariantLoci}
\alias{omitInvariantLoci}
\title{Calculate pairwise kinship for all pairwise comparisons across all individuals (only others) and loci (only invariant)}
\usage{
omitInvariantLoci(dataset)
}
\arguments{
\item{dataset}{uses result of convertBasePairstoCodes, which is the product of this code is used in classifyAllelePairsOthers and then determineAllAlleleMatchesOthers}
}
\value{
the supplied data frame with invariant loci (i.e., columns removed)
}
\description{
Calculates pairwise kinship across all individuals and loci and but excludes invariant loci
}
