% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGeneticData.R
\name{readGeneticData}
\alias{readGeneticData}
\title{Read in Genetic Data into R and Format Properly}
\usage{
readGeneticData(fileloc)
}
\arguments{
\item{fileloc}{The location of the data file, specified as a path (see documentation for \code{runGenets} and \code{runKinship} for information about \code{fileloc}).

Note that if using this function on its own outside of the \code{runGenets} or \code{runKinship} wrapper functions, you must specify the path to the genetics data file manually e.g.,

\code{readGeneticData(fileloc = paste0(getwd(),"/Data/", paste0(list.files(paste0(getwd(),"/Data/")))))}

if the genetics data are located (as they should be) in a working directory folder called \code{Data}.}
}
\description{
With the \code{fileloc} specified, one can import the initial data that's got paired alleles (e.g., C:G, A:T, etc.) along with a \code{Coral_ID} columns and any other columns like site name. This function names the first column \code{Coral_ID} and ensures that R interprets "T" as a character rather than the logical \code{TRUE}. The function also omits any potential white space (e.g., " C:T"). Lastly, \code{NA} values are converted to \verb{?} because we need to keep track of allele combos that couldn't be compared, and \code{NA} could be problematic when converting allele pairs to single letters with the \code{IUPAC} cipher.
}
